Fix QtInspector failing to locate shared library on Linux
As explained in the bugs section at http://linux.die.net/man/3/dlopen ,
dladdr(&symbol, &info) can return unexpected results if '&symbol'
ends up being resolved to the address of the thunk in the PLT
instead of the real symbol.

Use the suggested solution on that page to compile main.cpp
with -fPIC to avoid this issue.